Removed before setup and setup events from lifecycle. Optimized creating of lifecycle instance in component.
