+++ src/lib/setup/base.js
@@ -35,8 +35,14 @@ const shaderAlias = {
   rounded: 'radius',
 }
 
-export default (component) => {
+export default (component, name) => {
   Object.defineProperties(component.prototype, {
+    name: {
+      value: name,
+      writable: false,
+      enumerable: true,
+      configurable: false,
+    },
     focus: {
       value: function (e) {
         Focus.set(this, e)
